+++ opentelemetry-sdk/src/opentelemetry/sdk/_logs/_internal/export/__init__.py
@@ -33,7 +33,11 @@ from opentelemetry.sdk._logs import (
     ReadableLogRecord,
     ReadWriteLogRecord,
 )
-from opentelemetry.sdk._shared_internal import BatchProcessor, DuplicateFilter
+from opentelemetry.sdk._shared_internal import (
+    BatchProcessor,
+    DuplicateFilter,
+    RequestFilter,
+)
 from opentelemetry.sdk.environment_variables import (
     OTEL_BLRP_EXPORT_TIMEOUT,
     OTEL_BLRP_MAX_EXPORT_BATCH_SIZE,
@@ -51,6 +55,8 @@ _ENV_VAR_INT_VALUE_ERROR_MESSAGE = (
 )
 _logger = logging.getLogger(__name__)
 _logger.addFilter(DuplicateFilter())
+_wsgi_request_logger = logging.getLogger("django.request")
+_wsgi_request_logger.addFilter(RequestFilter())
 
 
 class LogRecordExportResult(enum.Enum):